Fix bug in processCatchBlock for catches from inliend frames
Summary: When we optimize away stack values in a catch block we cannot drop values which don't occur within the inner most frame within which the EndCatch is located. If we're throwing from an inlined function we may reenter an outer caller and therefore cannot assume that those stack locations will be killed by the unwinder.

Reviewed By: ricklavoie

Differential Revision: D29251922

fbshipit-source-id: 60bec9bc58c905450c0aed34f65a695800d35b54