Emit LdBindAddr inline
Summary:
LdBindAddr was loading an address from the data section.  Instead emit a
smashable load inline, and smash it to the correct address when bound.  We
should save a cache line load from the data section doing this.

We could intoduce a double word smashable load immediate instruction to save on
cache space, but for now keep it simple and repurpose the smashableMovq logic.

Reviewed By: jano

Differential Revision: D29264532

fbshipit-source-id: 91b08c9f1dab830f34a6fff24295dcf1e07594c3