Allow literal string optimization in global case
Summary: The global case already is supported. `pce->literal_data->matches` and `pcre_exec` are in a while loop that takes a `regex` and `start_offset` and checks for all matches at that offset. After the first match (incl. subpattern matches) is found, the subsequent searches are continued on from end of the last match. Since our pattern has no subpatterns (added an assert in case we ever expand this optimization and change this), we don't need to check multiple substrings at once.

Reviewed By: ricklavoie

Differential Revision: D29238870

fbshipit-source-id: 6fa4707367e128b7d3a9ec7f71829f1e2f523723