Fix BBC bug with SetOpL, math, and exceptions
Summary:
SetOpL with Add/Sub/Mul works by first converting the LHS to num and then doing the actual op. In a case like the following

```
  $a = null;
  $e = keyset[];
  try {
    $a += $e;
  } catch (Exception $e) {
    var_dump($a);
  }
```

This results in $a being `0` in the exception block. However, we didn't propogate this correctly in hhbbc. We assumed that the op had no effects whatsoever when an exception is thrown, meaning $a could be replaced with `null` in the catch block.

My first attempt at this fix was to update HHBBC to match the interpreter, but I've decided to instead just change HHVM.

Differential Revision: D29276917

fbshipit-source-id: ba34c0b0f4510af33445a2bd38c89cdcd551bda1