Skip compiling regex when literal data is used
Summary:
When the pattern is a literal, we can skip compiling it and use the literal_data to match instead.

In these cases, we also do not need to construct a name table because this is only used to add named subpatterns to the result, which literals inherently have none. (re: having named subpatterns requires `<name>` notation in the pattern.
```
> preg_match_all_with_matches('/(?P<id>[A-Za-z]+):25/', 'blue:25', inout $m, PREG_PATTERN_ORDER);
> var_dump($m);

dict(3) {
  [0]=>
  dict(1) {
    [0]=>
    string(7) "blue:25"
  }
  ["id"]=>
  dict(1) {
    [0]=>
    string(4) "blue"
  }
  [1]=>
  dict(1) {
    [0]=>
    string(4) "blue"
  }
}

```

 We also currently construct the name table with `pcre_fullinfo`, which requires `re` to be compiled, which this diff is aiming to avoid.

Cleaned up some of the loops of `preg_match` to make it easier to read / understand how implementation relates to spec from php manual.

Reviewed By: ricklavoie

Differential Revision: D29256359

fbshipit-source-id: 210b23cc454bf1a4eb7b4e1eab349d41ceeacec0