Replace folly::Optional/std::optional with HPHP::Optional
Summary:
Replace almost all usages of folly::Optional or std::optional with
HPHP::Optional instead. This change is mainly mechanical, except for a
few interface changes (reflecting folly::Optional incompatibility with
std::optional). The only places not changed are a few places
interacting with the rest of folly, expecting a folly::Optional, or
some extensions which I don't want to touch.

Reviewed By: mofarrell

Differential Revision: D29291295

fbshipit-source-id: c0701a6b1f0853a1b2ef28fa390a6005d523c258