avoid spawning leaser twice
Summary: We initialize the lease by creating an `env` variable carrying the input and leaser info used for typechecking on RE. Currently we create such an `env` when starting hh_server, and then create a new one when typechecking. This diff wraps the re `env` in `serverEnv` to thread it all the way through.

Reviewed By: bobrenjc93

Differential Revision: D29236390

fbshipit-source-id: 6386bd59557b3892754f0828b866cf9309141ec6