Implement checkProfile for OptPrologue
Summary:
The `reportRegion()` lambda in `checkProfile` assumed the TransKind is Optimize. The TransKind can now be OptPrologue, identified by the PrologueID instead of a RegionDesc.

`reportRegion()` triggers when there are missing block counters, or there is a mismatch between the opcode stored in the profile vs what we find in the optimized translation.

See [RFC] in test plan.

On V2 run, hhvm-flib-test-dbgo passes on all tests except 2 that were failing at the time.
Inspected all coredumps and have seen these all on another diff I have that is empty.

Reviewed By: paulbiss

Differential Revision: D29231873

fbshipit-source-id: 412dd9aaacbd12ca25f095b7a75fa50c201e4d5b