Allow is/as checks against varray/darray
Summary:
varray and darray typehints aren't QUITE vec and dict typehints, but we can at least make them agree in is/as position. Currently, we fatal on this usage, so it should be okay to start allowing it.

We don't need to update HHBBC. The only inference it makes about IsTypeStructC for varray/darray typehints is that the output is TBool.

Reviewed By: ricklavoie

Differential Revision: D29346005

fbshipit-source-id: a8a8263e61e905b2caabc2f0cbe94ec6e4d9b2e2