Disallowing memoizing objects in local contexts
Summary:
This diff disallows memoizing object and mixed types in contexts that don't have access to globals.

The task has a bunch of context on why we want to do this, but the short answer is to prevent data leaks caused by object types being memoized and then accessed again from just its key.

I will be landing this with a FIXME able error code for now, and then fixing our use cases for them later in www. There's only 65, so the number is relatively easy to fix

Reviewed By: viratyosin

Differential Revision: D28862056

fbshipit-source-id: 695b2d5e37a2ddc41b2fdece8b0075648351422a