Do not define order of SrcKeys
Summary:
There is no good reason to compare two SrcKeys and no good definition of
their total order. This will be even more so with the introduction of
FuncEntry.

Kill the total order and update IRBuilder::m_skToBlockMap to use jit::fast_map
instead of jit::flat_map. The flat map made sense when it was introduced, as
the code was using Offsets, inserted close to their order, and the translation
sizes were small. However, nowadays it is keyed on SrcKey, translation sizes
are large, there are regions that do not follow linear path and we skip across
functions due to inlining. Insert sorting is no longer appropriate.

Reviewed By: ricklavoie

Differential Revision: D29286996

fbshipit-source-id: f33ecc608406818ca72addb391e9941bba7351dc