Extra logic rules from Type and Method dep edge.
Summary:
Refactor `extract_logic_rules`, reuse the algorithm we wrote for "Extend", abstract out the part depending on edge data.

Created a family of classes to handle each type of the "edges", so far we only have the different in `lhs_parser` and `rule_writer`, the rest are sharing the same logic.

Reviewed By: jamesjwu

Differential Revision: D29241681

fbshipit-source-id: 357458378e0bf663655a3777dc52b74f1b5e0f0b