Replace pointer based serialization in prof data file
Summary:
The prof data file uses a pointer based serialization. That is, the
unique "id" used in the file for an object is just the object's memory
address. This works because the memory address is guaranteed to be
unique. However, this scheme falls apart if you want to
deserialization or append to the file across process restarts. If that
happens, the memory addresses of the same logical object changes, so
they may be serialized differently.

Since we want to try a "double jumpstart", we need to fix
this. Instead of using pointer based serialization, assign
incrementing ids. We regenerate this mapping when deserializing, and
persist it if we're going to append further data for the file. Since
the id isn't tied to the memory address, we can ensure a stable
mapping. As an added bonus, we can remove the "serialized" bit from
the various objects.

Differential Revision: D29322207

fbshipit-source-id: 7a292928bee8215efd589a5d56251875d507da0b