Profile and re-order RDS offsets
Summary:
During profiling translations, keep track of how often every item in
RDS is accessed. Use this profiling to optimally re-order RDS items
(by their hotness) and write that new ordering to the JIT profiling
file. During start-up, read out the ordering and use it to pre-assign
offsets to those hot RDS items. When actually binding a symbol, use
the pre-assigned slots if available.

The benefit of all this is that the hot RDS symbols are clustered
together, resulting in better cache utilization. This scheme works
best with double jumpstart, because we must restart before the
re-ordering can take effect.

Differential Revision: D29323165

fbshipit-source-id: 347694d4ad037fe5b332780e4ce25624047f1dfa