Implement "double jumpstart"
Summary:
vasm-block-counts can have a bad interaction with certain
optimizations. If an optimization requires a restart to take effect
(for example, if it requires reading from the jit profile file), them
it won't be present when vasm block counts is gathered. When we
restart after vasm block counts is done, the optimization kicks in,
but now the resultant code no longer matches the vasm block counts.

"Double jumpstart" is one approach to solve this. Right now, after
profiling, we immediately do RTA, then gather vasm block counts, then
restart. Instead, after profiling, we'll restart (without running
RTA). After restarting, we then run RTA, then gather vasm block
counts, then restart (again). This gives any of the above described
optimizations a chance to kick in before vasm block counts.

This implements the double restart scheme. It is gated on the
Jit.SerializeOptProfRestart option (defaults to true).

One complication with this scheme is that it's harder to coordinate
what "phase" the jumpstart should be in. This is done by extending the
"part" file extension scheme to account for multiple phases.

Modify the test runner to deal with the extra restart.

Differential Revision: D29322442

fbshipit-source-id: 2b0cf9732f51f3a9b88a124a7184b2cbcdcc63bf