Implement default values for abstract class constants
Summary:
This is a simple extension to class constants that has been enabled by previous work to resolve default values for abstract type and context constants. It admits syntax
```
abstract const int X = 3;
```
and the semantics are identical to default values for the other two kinds of constants. As with the other constants, the abstract with default compiles to the `isAbstract` HHAS tag, which is used to determine inheritance order in HHVM. As such, there are no runtime changes necessary, and only one HHBBC change to make sure the static analysis is correct. The rest of the diff is just enabling the syntax in the parser and threading the information through to the compiler / typing.

**Note:** Unlike concrete type and context constants, concrete class constants are still allowed to change in child classes. Banning that requires a separate effort; bottom line: semantics of constants are nearly converged but we're not quite there.

Reviewed By: losvald

Differential Revision: D29155489

fbshipit-source-id: e5658420801c8a913aef32557672eb38b7408aa8