Cache the folly::dynamic used to compute RepoOptions
Summary:
When hhvm calls into hackc to compile bytecode we pass a JSON blob for the RepoOptions associated with the request. For some reason JSON parsing in rust is horrendously slow, and so an LRU cache was added (as these options seldom change) to avoid completely wrecking frontend performance. Unfortunately the JSON blob is constructed from a folly::dynamic which is built on the fly for each new parse, and the hash map backing folly::dynamic has non-deterministic ordering in debug builds (but the order does not change once its constructed).

So while parse time was fixed by caching in our prod builds debug builds continued to suffer. This diff caches the folly::dynamic when the (already cached) RepoOptions are constructed, thus allowing us to hit in the cache.

Reviewed By: mofarrell

Differential Revision: D29410266

fbshipit-source-id: 6fdf230aa0a4ed6e3dd148380f5c58e08d21e915