add Diagnostic_pusher module
Summary:
Diagnostic_pusher tracks which errors have been sent to the LSP client in order to determine which errors to erase.

It is very similar to Diagnostic_subscription but more adapted to streaming errors.

For now, this module is not used but it is self-contained and already reviewable. It has its own unit test Diagnostic_pusher_test.

It will be enhanced later to track losing client and getting new ones.

The API should be well documented, so review .mli files before .ml files.

Differential Revision: D29304720

fbshipit-source-id: c42ace6a0759eb041f2e397ca98deb97a1116fa5