store persistent client in client provider
Summary:
Currently, the persistent client is stored in ServerEnv.env.persistent_client.
To be able to stream errors, we need to access the persistent client from Typing_check_service, which does not have access to the server env.

We could wire down the persistent client in there, but we would miss new clients that ServerMain.priority_client_interrupt_handler would get.

This diff stores the persistent client in a reference in ClientProvider, in addition to storing it in env.persistent_client. We'll clean up env.persistent_client in a later diff.

Persistent clients are created by ClientProvider.make_persistent, so we set the persistent client reference there.

Differential Revision: D29304718

fbshipit-source-id: 948abd6f8f3bd3a1bd3814d2aacaa59dc35ca649