Added function to emit Hack code on Type dependency.
Summary:
Added `add_parameter` to handle the "**Type**" dependency edge, where there is no "Extends" relationship, so the "**Type**" dependency was created by passing the Type as parameter to the other Type.

Currently, we are creating a dummy method inside each type to accept the parameter list, it produces the same dependency graph, but the code is not very realistic. We can improve this later on.

The dummy method is named as "dummy_{name}_method", if user provided the same name as their method, we are appending the underscore ("_") to the end until we get a non-conflict name.

Reviewed By: madgen

Differential Revision: D29313996

fbshipit-source-id: 03c8a2223fb841a5c792ec3ec2c76cff221ba7c1