Demote defaults for class constants to Unstable
Summary:
There's a bug in HHVM where I've conflated uninit with not having a value, so this breaks for class constants

```
interface I {
  abstract const int X = OtherClass::K;
  abstract const int Y;
}
```
Both of these pass isAbstractAndUninit() which means that a later constant in the initialization, possibly one without a default, will win. Demoting the syntax to fully Unstable in the meantime so that it doesn't compile.

Reviewed By: losvald

Differential Revision: D29447426

fbshipit-source-id: db71a5a015c946f426130bfe7768906aabdd7862