Refactoring Symbol Map Unit Test
Summary:
This is a refactor of the symbol map unit test that makes the code a bit more concise by taking advantage of some of the available gmock container matchers.  The intent of this change is not to change the underlying test conditions of any tests.

This affects a single unit test, which passes.  The other test failures are unrelated to this change.

Reviewed By: jthemphill

Differential Revision: D29157044

fbshipit-source-id: 3aeb4ab2a24a55818b800a2a2337a6b3e96e373f