Update srcs for a store even for srcs that admitSingleVal()
Summary:
In store-elim, when moving stores between blocks, we first check if a store is *available* at a certain point; ie it has happened on all paths to this point, and there is no interference between those stores and this point. In this example, a store is available In to B113. When we try to sink the store to B113, we skip updating the sources with the dest of the phi of B1888 and B1967 because both of the stores have sources that admit the same single val. This is incorrect because we end up with temps that necessarily do not strictly dominate their uses.
```
-------------------------------Assertion Message--------------------------------
src 't1537:InitNull' in '(8804) StStk<IRSPOff -28> t1:StkPtr, t1537:InitNull' came from '(8317) t1537:InitNull = DefLabel', which is not a DefConst and is not defined at this use site
-------------------------------Assertion Failure--------------------------------
hphp/runtime/vm/jit/check.cpp:214: bool HPHP::jit::checkCfg(const HPHP::jit::IRUnit&): assertion `locally_defined || strictly_dominates' failed.
-------------------------------------IRUnit-------------------------------------
```
```
 B113: [profCount=118] (preds B1967 B1888)
    --- bc ReadStateLogger::genLogEvents(id 1711356672)3671, fp 0, spOff 7  [profTrans=1666077]
    AssertRATStk 0 ?Int
           t1538:InitNull = phi t1537:InitNull@B1967, t1532:InitNull@B1888 <--- the phi dest
           t1540:Dict=StructDict<"vpv_id","image_loading_state","mf_story_key","recent_vpvs_graphql_filter"> = phi t1191:Dict=StructDict<"vpv_id","image_loading_state","mf_story_key","recent_vpvs_graphql_filter">B1967, t1158:Dict=StructDict<"vpv_id","image_loading_state","mf_story_key","recent_vpvs_graphql_filter">B1888
    (8318) t1538:InitNull, t1540:Dict=StructDict<"vpv_id","image_loading_state","mf_story_key","recent_vpvs_graphql_filter"> = DefLabel
    (8319) t1539:InitNull = CheckType<InitNull> t1538:InitNull -> B114

...
 B114: [profCount=35] (preds B113)
    --- bc ReadStateLogger::genLogEvents(id 1711356672)3671, fp 0, spOff 8  [profTrans=1729573]
    AssertRATStk 0 ?Int
    (8321) t1541:Bottom = CheckType<Int> t1538:InitNull -> B2051<Unlikely>
...
 B2051<Unlikely>: [profCount=35]<Unlikely> (preds B114)
...

    --- bc ReadStateLogger::genLogEvents(id 1711356672)3656, fp 0, spOff 6  [profTrans=858136]
    FCallClsMethodD <SkipRepack> 2 1 "" - "" "FeedTrackingUtils" "FeedTrackingUtils" "getIntIfNotNull"
    (8804) StStk<IRSPOff -28> t1:StkPtr, t1537:InitNull <-- the store
```

{F628088917}

Reviewed By: mofarrell

Differential Revision: D29406550

fbshipit-source-id: 5dfed79e2752cbbcfffc9776e5be0dbade45ef63