add _Str\vsprintf_l
Summary:
So, this took a while, as the variadic form (sprintf) raised some
issues with the original plan.

We were going to add an optional locale parameter to all `Str\` functions;
that's not possible for variadic ones.

New plan is to add `_l` variants of everything, with mandatory (but
potentially nullable) locale parameters. TBC in user-facing API: first arg, or last mandatory arg.

This avoids this issue, and potentially allows us to optimize the common
case more (or at least more easily) in the future.

The locale parameter placement in this private builtins feels weird, but:
- it's a private builtin so doesn't really matter
- `sprintf_l` and friends actually exist on BSD; matching them where
  possible seems like the least surprising thing to do.

Reviewed By: oulgen

Differential Revision: D29315004

fbshipit-source-id: b088c9cf59280150f9ba9d851292e173c4de2074