5/5 log anomalous ndecl_file_error_if_already_bound
Summary:
I seek to understand the code in Naming_global. This diff adds docblocks for two functions based on my reading of how the functions are intended to be used:
```
(** This function "declares" top-level names, i.e. adds them into the naming-table provider.
The caller is expected to ensure that all names from the specified file have already been removed
from the naming table prior to calling this function. It is also expected provide "full" positions
in its FileInfo.t. If any symbol from the FileInfo.t is already defined in a different file
(with or without the same case), that other file is deemed to have the "canonical" definition
of the symbol; this function leaves the naming-table for that name as it is, and produces an error,
and also returns the canonicals filename. If any symbol from the FileInfo.t is defined twice within the FileInfo.t
(with or without the same case), then the first occurrence is deemed to be the "canonical" definition
of the symbol; this function leaves the first occurrence in the naming-table, and produces an error
on the second, and returns this filename. *)
val ndecl_file_error_if_already_bound :
  Provider_context.t ->
  Relative_path.t ->
  FileInfo.t ->
  Errors.t * Relative_path.Set.t

(** This function "declares" top-level names, i.e. adds them into the naming-table provider.
This caller is expected to ensure that there are no naming-collisons and no case-insensitive naming collisions. *)
val ndecl_file_skip_if_already_bound :
  Provider_context.t -> Relative_path.t -> FileInfo.t -> unit
```

It also adds runtime invariant-checking to back up the expectations I wrote down for ndecl_file_error_if_already_bound:
* The caller is expected to ensure that all names from the specified file have already been removed
from the naming table prior to calling this function
* It is also expected provide "full" positions in its FileInfo.t

This diff doesn't fail or assert or anything. If the expectations are violated, it merely logs telemetry. If I examine that telemetry in a week or two and find no violations, then I'll know I've correctly discerned the nature of this code and its callers! This will help me refactor it further in future.

Reviewed By: CatherineGasnier

Differential Revision: D29365375

fbshipit-source-id: 10065102fff4b43bb47b3cf8a25d06d87f966939