hh_single_type_check call out side-effects of naming+decling
Summary:
In hh_single_type_check and hh_single_complete, the function `parse_name_and_decl` does a lot under the hood. In this diff I've added comments to explain what the functions do, and moved some of the functionality around a little.
* **parse_and_name**. "This is an almost-pure function which returns what we get out of parsing. The only side-effect it has is on the global errors list." *(prior to this diff, it also updated the global reverse naming table and gathered duplicate-name errors).*
* **parse_name_and_decl**. "This function is used for gathering naming and parsing errors, and the side-effect of updating the global reverse naming table (and picking up duplicate-name errors along the way), and for the side effect of updating the decl heap (and picking up decling errors along the way)."
* **add_decls_to_heap**. "This function is used solely for its side-effect of putting decls into shared-mem". *(prior to this diff, this function didn't exist, and people instead used parse_name_and_decl with all its extra behavior too)*.
* **get_shallow_decls**. "This function doesn't have side-effects. Its sole job is to return shallow decls.". *(prior to this diff, this function also had the side effect of updating the global reverse naming table, and gathering then ignoring errors about duplicate names)*.

Why am I writing this diff now? It's because I'm trying to clamp down on `Naming_global.ndecl_file_error_if_already_bound`. That is the function which updates the global reverse naming table and gathers up duplicate-name errors.

Prior to this diff, a lot of hh_single_type_check and hh_single_complete scenarios would add a file's contents to the global reverse naming table twice. This wasn't on purpose. It was just an accident along their way to updating the global search index, or getting shallow decls, or other things. They were inadvertently saved by the fact that "add file's contents to global reverse naming table" has a hack to make it idempotent. (in particular, if you add a name, but that name was already defined, then it should report "error name is already bound"; the idempotency hack is that if the name was already defined in the exact same file and line and column as you're defining it now then we suppress the error).

I'm want to get rid of the hack. I aim for a principled implementation of "add file's contents into the global reverse naming table" which simply adds it, and a pre-existing name clash is ALWAYS an indication of error.

In order for me to get rid of the hack, I first have to change hh_single_type_check and hh_single_complete so that they no longer depend upon the hack. That's what this diff does.

Reviewed By: CatherineGasnier

Differential Revision: D29395017

fbshipit-source-id: 7066149bc3978ade3e4dd347c7aa08360bdd0378