Treadmill away APC bespoke arrays
Summary:
We use compare_exchange_strong to ensure that only one thread computes the final layout of a given APC bespoke array. Have that thread treadmill the vanilla and logging arrays after doing so.

To make this change work, we have to modify the destructor of an APC bespoke: if the final layout has been computed, the destructor does not destroy the vanilla and logging arrays. We also have to write back this array to the outer APCTypedValue, because the value there is used for stats computations.

Reviewed By: colavitam

Differential Revision: D29479256

fbshipit-source-id: 1106efdc1c0b6d7e8b569d1b323c23c3cfbc566f