Restore "never modified" static property optimization
Summary:
We used to have an optimization for static properties which
are never modified (beyond their initial value). For such properties,
we can bind the RDS handle into persistent. If the class contains
nothing but such properties, we can skip per-request sprop
initialization entirely. This broke at some point due to a HHBBC
refactor. This restores the optimization.

Reviewed By: paulbiss

Differential Revision: D29492789

fbshipit-source-id: 3733421eae4b3056c07afc2a598982b579ac97dd