Fix coeffect constants that are not runtime enforced
Summary:
Previously, unrecognized coeffects would result as pure in the runtime, i.e.
```
const ctx T = [codegen];
```
would result in
```
const ctx T = [];
```
in the runtime. This diff makes it so that it now results as
```
const ctx T = [defaults];
```
This is done by transferring the name of the coeffect to the runtime and letting runtime make it defaults. This change will also allow adding reflection in the future.

Reviewed By: arnaudvenet

Differential Revision: D29550306

fbshipit-source-id: d77e552b890729a32f1b774292445f7af580b494