longer timeouts for tests
Summary:
Sunday July 4th had some test failures https://fb.workplace.com/groups/hacklang.release/posts/1046751659194416

The failure is that `--retries 20` parameter has run out without the monitor yet responding to hh_client.

This failure has happened one other time in the past fortnight:
https://www.internalfb.com/intern/test/844424943247988?ref_report_id=0

https://pxl.cl/1LdZj

I think this is just an inevitable fact of life. Especially on a busy machine that's running lots of tests concurrently. With this diff I'm bumping up the timeout to 3mins. This will mean that we take longer to get signal on broken tests. But the reality is (1) we don't run tests locally, (2) we don't break these tests often.

There are a few future directions that we could take:
1. have all sandcastle tests pass a much longer timeout, and use a shorter timeout for dev-machines. (But, I don't really like having different setup for tests! it means you can't reproduce them as well!)
2. treat all such things where hh_server is too busy to respond as "skipped" rather than "failed". We already do that for the LSP tests.
3. Engineer hh_server so every test takes just a single processor, and so won't run into troubles with `buck test` parallelism. That's what we're doing in the re-architecture.
4. Mess about with buck test TARGETS files so as to reduce parallelism and make everything take longer. That'd be a downside.

So with this diff I've opted for the least impactful way of eliminating signal noise.

Reviewed By: shayne-fletcher

Differential Revision: D29543999

fbshipit-source-id: c2014d859074443084de39bd09f5dd384195f331