Added function to emit Hack code on invoking a method.
Summary:
Added `add_invoke` to handle the "**Method**" dependency edge, where there is **no "Extends"** relationship, so the "**Method**" dependency was created by passing the Type A with a Method foo as parameter to the other Type B, and invoke the A::foo inside B.

And only class method has a body can invoke methods inside, we don't have this `add_invoke` for the interface.

Currently, we are creating a dummy method inside each type to accept the parameter list, and we invoke all of the non-override methods inside this dummy method. It produces the same dependency graph, but the code is not very realistic. We can improve this later on.

In `on_model` callback, we are adding a new pass to handle any predicate with arity of 3. All of the relying information should be ready before third pass, e.g, has valid class/interface been created using `add_class/interface`, has method been added using `add_method`, etc.

Reviewed By: jamesjwu

Differential Revision: D29314693

fbshipit-source-id: 642ba715c05344bad044334c72cf00de25636ecc