Use nearby RIP-relative addressable placeholder address
Summary:
ldbindretaddr and leav vasm instructions use a placeholder RIP-relative
address, which is bound later by vasm emit. This placeholder address needs
to be close enough to the code to be emittable, so let's use the current code
address. The previous constant was within 2GB of almost whole low memory, but
HHVM may emit code outside of low memory in certain configurations.

Reviewed By: ricklavoie, mofarrell

Differential Revision: D29565501

fbshipit-source-id: 22f245613fef8c1a2e267f84ba3700cb84b5a31d