Extract all class dependencies from type hints
Summary:
Currently only class names appearing in `Happly` are added as dependencies. This means, for instance, that classes whose type contstants appear as bounds on type parameters are not extracted, e.g. T94596648.

This diff causes all relevant class names to be added as dependencies wherever they occur in a hint

Reviewed By: madgen

Differential Revision: D29589625

fbshipit-source-id: a97728f9dc9d5a94b94b3d129be0bfd446b83d2e