Add test for access type constant from unresolved type
Summary:
There is a typing issue in expression trees related to function calls.

If a function being called within an Expression Tree has type constant constraints, due to the way desugaring works, we throw an error being unable to project the type constant off an unresolved type.

This happens when we type the desugared code (because we only get access to the function pointer), as opposed to the virtualized code (which is able to resolve the types based on the arguments being passed in).

This diff adds the following test to the corpus, just as a way of keeping track of it.

```
abstract class MyBox {
  abstract const type TInner;
}

class IntBox extends MyBox {
  const type TInner = ExampleInt;
}

async function setState<T as MyBox, TVal>(
  ExampleContext $_visitor,
): Awaitable<ExprTree<Code, Code::TAst, (function(T, TVal): void)>>
where
  TVal = T::TInner {
  throw new \Exception();
}

function test(ExprTree<Code, Code::TAst, IntBox> $x): void {
  Code`setState(${$x}, 1)`;
}
```

As one can see, the call itself, during virtualization, will be able to be properly typed and the type variables can be properly resolved. However, note that when we desugar this, we get the following:

```
function test(ExprTree<Code, Code::TAst, IntBox> $x): void {
  (
    () ==> {
      $_splice0 = $x;
      return Code::makeTree(
        darray[
          "path" => __FILE__,
          "start_line" => 24,
          "end_line" => 24,
          "start_column" => 7,
          "end_column" => 25,
        ],
        darray[
          "splices" => dict["\$_splice0" => $_splice0],
          "functions" => vec[setState<>],
          "static_methods" => vec[],
        ],
        (
          (Code $_v) ==> {
            return $_v->visitCall(
              darray[
                "path" => __FILE__,
                "start_line" => 24,
                "end_line" => 24,
                "start_column" => 7,
                "end_column" => 25,
              ],
              $_v->visitGlobalFunction(
                darray[
                  "path" => __FILE__,
                  "start_line" => 24,
                  "end_line" => 24,
                  "start_column" => 7,
                  "end_column" => 25,
                ],
                setState<>,
...
...
```

It is the function pointers that trigger the error, because at the time of resolution, the type variables are unable to be resolved because they are not paired with the arguments the function is supposed to be called with.

Reviewed By: Wilfred

Differential Revision: D29476411

fbshipit-source-id: 017c626cf3c04ca755521b94b971de29ffc91a13