Compress LowStringPtr immediates
Summary:
Now that we no longer use lowptr HHBBC, we store string bytecode immediates in 8 bytes rather than 4. However, most strings are still low pointers, so we can take advantage of this fact where possible.

Because StringData is aligned to 8 bytes, we can use the low bit to distinguish between the two cases.

Reviewed By: ricklavoie

Differential Revision: D29592630

fbshipit-source-id: 86151d54fc88b24c470bc7f62fc741ea80bf793e