Compress bytecode in FuncAnalysis results
Summary:
We already have the ability to compress an arbitrary vector of Bytecode. Apply compression to the replacedBcs results in a FuncAnalysis block update list.

Doing so is simple and type-safe; we can compress the bytecode when we store it to the analysis result, and expand it when we read it for optimization. Because this change compiles, we know that we aren't reading any of the compressed data in the middle.

Reviewed By: ricklavoie

Differential Revision: D29599756

fbshipit-source-id: 2d0786d1b50b52ec3e25d91b62c77a65b6063048