extend negation types to classes
Summary:
Previously, negations only applied to primitives, now they apply to classes also. The negation type represents the type of a variable where an `is` check that returns false. Since `is C<_>` doesn't allow type arguments, the type of a negated class has no arguments, and the intended semantics is that if `e : negation C`, then forall types `t`, the value of `e` cannot be a `C<t>`.

This is a straightforward generalisation of negation for primitives, except that it has a complicated interaction with the way type arguments are handled when refining a generic class. Consider the following function,
```
function f(mixed $v) : void {
  if ($v is Vector<_>) {
    // $v : Vector<Tv#1>
  }
  else {
    // $v : not Vector
  }
  // $v : Vector<Tv#1> | not Vector
}
```

We want to simplify the union back to `mixed` to keep types compact, and match current behaviour, but the fact that the contents of the vector are `Tv#1` (a fresh type variable), leaks from the scope of the block where the refinement happened (although we don't take any advantage of that). `Vector<Tv#1> | not Vector` is not equal to `mixed` because it doesn't include any vectors that aren't `Tv#1`. However, here we can soundly give `$v` a super-type here. To accomplish this, the `union` function gets an optional parameter directing it to do this simplification.

This applies slightly too often, and also simplify to `mixed` in the following case.
```
function f(mixed $v) : void {
  if ($v is Vector<_>) {
    // $v : Vector<Tv#1>
    $v = Vector<int>{ 1 };
    // $v : Vector<int>
  }
  else {
    // $v : not Vector
  }
  // $v : Vector<int> | not Vector
}
```
However this is the current behaviour, so it isn't a problem in practice.

Reviewed By: CatherineGasnier

Differential Revision: D29562125

fbshipit-source-id: 7ce8631c55fdd7ee2118189e726446100ce368e8