We give the solver complete flexibility to choose from override or invoke a method.
Summary:
We are giving complete flexibility for the "automated reasoner" to choose from override or invoke a **method**. So that our code synthesis can be more creative.

Detailed analysis provided in the comment, so the future maintainer knows how to make trade-off between performance and creativity.

Renamed the variable name, given a meaningful name other than `X` `Y` `Z`.

Reviewed By: jamesjwu

Differential Revision: D29396989

fbshipit-source-id: e36c910d5eaf69beea9b2bc0b0ab39216077c3c9