Implement readonly check (instead of readonly convert)
Summary:
NOTE: This uses a non-mutating visitor (unlike the previous version). Making it mutating appears to be a pain because the rest of aast_parser doesn't pass down mutable programs... and changing it to be mutable seems to be some work (that I think is outside of the scope of this diff). If we want to go down this route then I can stack this on top of another diff that does that work.

NOTE: The original code in closure_convert seems to have been removed. It's possible it's because of the `by_ref` changes - not sure - I didn't look into it in too much depth. I based this off of my original diff...

Reviewed By: jamesjwu

Differential Revision: D29320968

fbshipit-source-id: cb9d1923528e4711aab304b3e19ca5013ad1eda5