autocompleteService supports builtins
Summary:
When you write "function f():string" and press ENTER, autocomplete doesn't know about the "string" type. It does however know about the "StringBuilder" type and suggests that (depending on the vagaries of precisely when the search for global symbols was made, and how quickly it got back). A user described this behavior as *maddening*.

https://www.internalfb.com/intern/px/p/1JKX4

With this diff, we now provide builtin types as suggestions in type positions.

https://www.internalfb.com/intern/px/p/1JKX2

I also took this opportunity to provide some docblocks for the builtin types. I got them from https://docs.hhvm.com/hack/built-in-types.

Note: the existing autocompleteService only distinguishes three kinds of places that expect types: (1) where a trait is expected, (2) after the "new" operator, (3) everywhere else. I've put the builtin types in category 3. That means for instance that we now suggest the builtin type "void" in function return type, but also in places where it doesn't apply like function parameters and the "is" operator. That's ideal, but I don't think it's worth going to extra lengths to fix.

Note: this diff might actually cause unpleasant UX. The deal is that sometimes (often?) the user does typeahead quicker than Hack can look up a list of all global symbols, and also Hack avoids doing a search for globally defined symbols unless the user has typed at least a few characters because there will be so many. But with this, the builtins will always be offered, even if we haven't look for user-defined symbols. I don't yet have a judgment whether this will be nice or not.

My plan is just to release it and gather feedback.

Reviewed By: Wilfred

Differential Revision: D28777576

fbshipit-source-id: 5bc9643e3ec9be6453798df2d715a444569fa45f